Registry backend memlog (#8131)
The memlog package implements a backend for the new registry store.

The memlog backend keeps all key/value pairs of an open store in memory. Changes made inside transactions are written to the disk-store, which is kept in sync with the in-memory data. A transaction fails and is rolled back if the disk-store reports an error while committing the transaction. When a store is opened, the disk-store is opened first and used to load a consistent and valid state into memory. On close the disk-store will be closed (transaction log file) and the in-memory state is dropped.

A memlog store supports multiple concurrent readers and one concurrent writer. All active transaction are isolated from each other. Readers can not access any live updates available in the write transaction. A write transaction will see all updates made within the current transaction.

The disk-store used by a memlog store is kept in a directory. The directory name matches the store name. The directory contains:
- data file(s)
- transaction log file
- symlink to most recent data file (active.json)
- meta-file with store version

A data file is fully compatible to the current registry format used by filebeat. It is an array of JSON objects. The data file name contains the transaction ID it was generated for: `<transaction id>.json`. The data file contains a full snapshot of the complete store it's state.

The `active.json` file points to the most recent valid data file containing a complete snapshot of the store. Some users write custom script to read the filebeat registry in order to check the amount of bytes processed by filebeat. These information are used for monitoring purposes, but also to trigger some cleanup. Users will have to upgrade their scripts by reading the ${path.data}/registry/filebeat/active.json file instead.

The transaction log stores operations applied to the in-memory state. By loading the data file and replaying the transaction log, a valid state of the in-memory store can be reconstructed.
The transaction log is a stream of json action-objects-pairs. Each transaction block has this form:

```
    {"action": "begin"}{"id": <transaction id>}
    ...
    {"action": "set"}{"k": "<key>", "v": {...}}
    {"action": "update"}{"k": "<key>", "v": {...}}
    ...
    {"action": "commit"}{"id": <transaction id>} // must match id of 'begin'
```

When loading the transaction log, all complete transactions with “transaction id > data file transaction id” will be applied. If a transaction block is incomplete or if a parsing error occurs, then the active transaction is rolled back and the transaction log is marked as corrupt. In this case, a new data file will be generated on the next write transaction and the transaction log will be truncated.
On error, all additional updates (transactions) made in an earlier beats instance to the store are lost. Yet we did load a valid state that has been correct in the past. This gives beats a chance to continue from an old state, which can still be updated and processed.

Memlog has 2 commit modes:
- transaction log commit
- snapshot commit
 
The transaction log commit mode just appends a new transaction block to the transaction log file. The is no additional fsync on the transaction log. If the machine/disk loses power or OS crashes before having had a chance to flush IO caches to disk, then the transaction log will be corrupted (out of date). Still, the loading process guarantees we will have some valid state of the store on restart. Here lies a trade-off between creating back-pressure inside the beat (overall throughput) and always having the latest updates available on restarts. Running tests on my machine (SSD), a transaction log only test run requires 0.1s (no fsync), whereas tests creating checkpoints which need fsync, even on small files take like 7.5s.

The snapshot commit mode is a little more complex. It resembles the old approach of writing a new complate registry file containing a snapshot of all accumulated file states:
1. copy the in memory state and apply all updates (keep old state in case the transaction fails)
2. write new state to temporary data file -> Fsync + close
3. rename temporary data file to <transaction id>.json
4. truncate transaction log
5. create new temporary active.json.tmp, pointing to the new data file
5. rename active.json.tmp to active.json (rename guarantees there is always one active.json)
6. fsync parent directory (often required if FS has no journaling support)
7. remove old data files.

Once step 3 succeeds and the new file is still visible after a restart (FS meta data updated before crash), then the transaction has been complete and the beat will restart with the new state.

Note:
Snapshot commits are quite expensive (amount of data to be serialized + fsync) and can very much influence filebeat throughput. Additional back pressure due to registry updates is known to influence overall throughput. We might look for other criteria as well:
time-based: snapshot commit should be run only every 5 minutes, so user scripts see some state updates.
Allow transaction log to grow quite large if store contains only 1 or 2 entries, but throughput shall be tuned.